requirements.txt
aiohttp             3.7.4
asgiref             3.3.4
async-timeout       3.0.1
attrs               21.2.0
beautifulsoup4      4.9.3
bs4                 0.0.1
certifi             2020.12.5
chardet             4.0.0
crawler             0.0.2
Django              2.2.10
django-simpleui     2.1
djangorestframework 3.9.1
greenlet            1.1.0
idna                2.10
idna-ssl            1.1.0
importlib-metadata  4.0.1
lxml                4.6.3
multidict           5.1.0